Add files via upload
Initial implementation. This is a quick demo of how one can call the ChatGPT API from a simple web page. The idea is to read all the Word (.docx) files from the chosen folder, concatenate all the text content in those files, and pass the content as a param to the ChatGPT API. You can choose from a few ChatCPT models; not all of them work but that will be rectified soon.